Fix clippy warnings
- Simplify identical if branches in field_widths.rs
- Replace unnecessary_unwrap with if-let pattern in process.rs
- Use saturating_sub for manual arithmetic checks (auto-fixed)
- Use char array for split pattern (auto-fixed)
- Remove unnecessary cast (auto-fixed)
- Simplify identical else-if branches in main_panel.rs
- Rename LastColorElement to Last to avoid enum_variant_names warning
- Add #[allow(clippy::too_many_arguments)] to UI modules where
  drawing functions naturally require many parameters

+++ src/ui/screen_manager.rs
@@ -239,7 +239,7 @@ impl ScreenManager {
 
                 if x + indicator_width < max_x {
                     let _ = addstr(pause_indicator);
-                    x += indicator_width as i32;
+                    x += indicator_width;
                 }
             }
